requirements.txt
youtube-transcript-api
bilibili-api-python
openai
python-dotenv
yt-dlp
notion-client
